day15_capstone/aim4_writer/requirements.txt
git+https://github.com/hypercycle-development/pyhypercycle-aim.git#egg=pyhypercycle_aim
groq==0.9.0
httpx==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0
markdown==3.5.1